Refactor cli commands and some imports
