+++ README.rst
@@ -19,7 +19,7 @@ A set of tools to provide a control plane for managing the lifecycle of Dask clu
 
 .. code-block:: bash
 
-    $ daskctl list
+    $ daskctl cluster list
     NAME               ADDRESS               TYPE          WORKERS  THREADS  MEMORY    CREATED   STATUS
     proxycluster-8786  tcp://localhost:8786  ProxyCluster        4       12  17.18 GB  Just now  Running
