+++ dask_ctl/__init__.py
@@ -6,10 +6,12 @@ del get_versions
 from .discovery import (
     discover_cluster_names,
     discover_clusters,
-    list_clusters,
     list_discovery_methods,
+)
+from .lifecycle import (
     get_cluster,
     scale_cluster,
     delete_cluster,
+    list_clusters,
 )
 from .proxy import ProxyCluster
