+++ dask_ctl/cli.py
@@ -1,24 +1,21 @@
 import datetime
 
 import click
-from tornado.ioloop import IOLoop
 
 from dask.utils import format_bytes, format_time_ago
-from distributed.cli.utils import check_python_3, install_signal_handlers
+from distributed.cli.utils import check_python_3
 
 from . import __version__
-from .utils import format_table
+from .utils import loop, format_table
 from .discovery import (
     discover_clusters,
     discover_cluster_names,
     list_discovery_methods,
+)
+from .lifecycle import (
     scale_cluster,
     delete_cluster,
 )
-
-
-loop = IOLoop.current()
-install_signal_handlers(loop)
 
 
 def autocomplete_cluster_names(ctx, args, incomplete):
@@ -38,7 +35,13 @@ def cli():
     pass
 
 
-@cli.command()
+@cli.group()
+def cluster():
+    """Cluster commands."""
+    pass
+
+
+@cluster.command()
 @click.argument("discovery", type=str, required=False)
 def list(discovery=None):
     """List Dask clusters.
@@ -97,27 +100,7 @@ def list(discovery=None):
     loop.run_sync(_list)
 
 
-@cli.command()
-def list_discovery():
-    """List installed discovery methods.
-
-    Dask clusters can be created by many different packages. Each package has the option
-    to register a method to discover clusters it creates. This command lists all discovery
-    methods registered on your system.
-
-    """
-
-    async def _list_discovery():
-        dm = list_discovery_methods()
-        format_output(
-            ["name", "package", "version", "path"],
-            [[m, dm[m]["package"], dm[m]["version"], dm[m]["path"]] for m in dm],
-        )
-
-    loop.run_sync(_list_discovery)
-
-
-@cli.command()
+@cluster.command()
 @click.argument("name", autocompletion=autocomplete_cluster_names)
 @click.argument("n-workers", type=int)
 def scale(name, n_workers):
@@ -137,7 +120,7 @@ def scale(name, n_workers):
         click.echo(f"Scaled cluster {name} to {n_workers} workers.")
 
 
-@cli.command()
+@cluster.command()
 @click.argument("name", autocompletion=autocomplete_cluster_names)
 def delete(
     name,
@@ -154,6 +137,32 @@ def delete(
         click.echo(e)
     else:
         click.echo(f"Deleted cluster {name}.")
+
+
+@cli.group()
+def discovery():
+    """Discovery commands."""
+    pass
+
+
+@discovery.command(name="list")
+def list_discovery():
+    """List installed discovery methods.
+
+    Dask clusters can be created by many different packages. Each package has the option
+    to register a method to discover clusters it creates. This command lists all discovery
+    methods registered on your system.
+
+    """
+
+    async def _list_discovery():
+        dm = list_discovery_methods()
+        format_output(
+            ["name", "package", "version", "path"],
+            [[m, dm[m]["package"], dm[m]["version"], dm[m]["path"]] for m in dm],
+        )
+
+    loop.run_sync(_list_discovery)
 
 
 @cli.command()
