+++ dask_ctl/tests/test_cli.py
@@ -4,12 +4,12 @@ from dask_ctl.cli import autocomplete_cluster_names
 
 
 def test_list_discovery():
-    assert b"proxycluster" in check_output(["daskctl", "list-discovery"])
+    assert b"proxycluster" in check_output(["daskctl", "discovery", "list"])
 
 
 def test_list():
     with LocalCluster(name="testcluster", scheduler_port=8786) as _:
-        output = check_output(["daskctl", "list"])
+        output = check_output(["daskctl", "cluster", "list"])
         assert b"ProxyCluster" in output
         assert b"Running" in output
